profile_file: fix array reads after end of data array
Issue was triggered by profile interpolation requested *exactly*
for the right endpoint of the profile provided via 'profile_file'.
In this case, function ProfileFile::value was processing
random data read from after the end of the data array 'xdat'
and 'ydat'.

Debugging started from crashes during the particle sorting process.
The issue fixed in this commit caused all particles in the
slice on the right endpoint of the profile x range to have
invalid values such as theta=-nan.