for now disabled display of "wall clock time" at end of program execution
Reason: Syscall clock() gives the CPU time spent by the process, not the wall clock time. For example, if your process is sleep()-ing for a longer time, this time is not included (as one would expect for a wall clock time display).
Therefore temporarily disabled the output until this is fixed.