Example3 readme
I believe this is a typo.

Also, as a newcomer, I have a hard time understanding the whole paragraph and the following.

E.g.: "The given example requests an integration step size of 45 mm, which corresponds to three undulator periods. This means that after one single integration step the field is pushed three wavelength." Why is the field pushed 3 wavelengths and not 3 undulator periods? (Sorry, I don't see the connection of the argument with delz.)

Is there a rule of thumb on how to choose 'delz' and 'sampel'?