CI: simple analysis of test simulation
This commit adds simple analysis of the output of the steady-state test simulation. If the result of the steady-state test sim is as expected, several key parts of the simulation code should work as expected.

Currently the test acceptance criterion is very relaxed, allowing a 30-percent relative power deviation from a hard-coded test value at the end of the beamline. The test acceptance criterion can be adjusted in `TEST_power.py`. If steady-state example1 is modified, one can run the script on the command line and make parameter adjustments.

In file `run_tests.sh`, the test using the final power from simulation output can be disabled by uncommenting the `exit 0` line.

This commit is equivalent to commits 61c94ab->bbed450 from CL's git repo, but with a different base commit.